removed nom from blank line
