+++ src/parse/parser_utils/predicates.rs
@@ -8,7 +8,10 @@ pub fn is_space_or_tab(c: char) -> bool {
 
 /// Returns whether the input is a blank line or not.
 pub fn is_blank_line(input: &str) -> bool {
-    BlankLine::parse_line(input).is_ok()
+    match BlankLine::parse_line(input) {
+        Ok((remaining, _)) => remaining.is_empty(),
+        Err(_) => false,
+    }
 }
 
 #[cfg(test)]
@@ -40,20 +43,43 @@ mod test {
         use super::*;
 
         #[test]
-        fn should_return_false_for_empty_string() {
+        fn should_return_false_with_empty_string() {
             assert!(!is_blank_line(""));
         }
 
         #[test]
-        fn should_return_false_if_contains_non_blank_characters() {
-            assert!(!is_blank_line(" \tword\n"));
+        fn should_return_false_for_string_with_one_none_whitespace_character() {
+            assert!(!is_blank_line(" \ta\n"));
         }
 
         #[test]
-        fn should_return_true_for_blank_line() {
-            assert!(is_blank_line(" \n"));
-            assert!(is_blank_line("\t\n"));
-            assert!(is_blank_line(" \t\n"));
+        fn should_return_false_for_2_blank_lines() {
+            assert!(!is_blank_line("\n\n"));
+        }
+
+        #[test]
+        fn should_return_false_for_a_blank_line_followed_by_a_character() {
+            assert!(!is_blank_line("\nabc"));
+        }
+
+        #[test]
+        fn should_return_true_with_space() {
+            assert!(is_blank_line(" "));
+        }
+
+        #[test]
+        fn should_return_true_with_tab() {
+            assert!(is_blank_line("\t"));
+        }
+
+        #[test]
+        fn should_return_true_for_carriage_return() {
+            assert!(is_blank_line("\r\n"));
+        }
+
+        #[test]
+        fn shoud_return_true_for_newline() {
+            assert!(is_blank_line("\n"));
         }
     }
 }
